feat(task): add configurable ssl context to amqp task
