+++ task-impl/src/main/java/com/chutneytesting/task/amqp/ConnectionFactoryFactory.java
@@ -1,9 +1,12 @@
 package com.chutneytesting.task.amqp;
 
+import static com.chutneytesting.task.common.SecurityUtils.buildSslContext;
+
 import com.chutneytesting.task.spi.injectable.Target;
 import com.rabbitmq.client.ConnectionFactory;
 import java.net.URISyntaxException;
 import java.security.GeneralSecurityException;
+import javax.net.ssl.SSLContext;
 
 public class ConnectionFactoryFactory {
 
@@ -11,6 +14,21 @@ public class ConnectionFactoryFactory {
         ConnectionFactory connectionFactory = new ConnectionFactory();
         try {
             connectionFactory.setUri(target.uri());
+
+            if ("amqps".equalsIgnoreCase(target.uri().getScheme())) {
+                SSLContext sslContext = buildSslContext(target).build();
+                connectionFactory.useSslProtocol(sslContext);
+            }
+
+            target.property("sslProtocol")
+                .ifPresent(protocol -> {
+                    try {
+                        connectionFactory.useSslProtocol(protocol);
+                    } catch (GeneralSecurityException e) {
+                        throw new IllegalArgumentException(e.getMessage(), e);
+                    }
+                });
+
         } catch (URISyntaxException | GeneralSecurityException e) {
             throw new IllegalArgumentException(e.getMessage(), e);
         }
